==> 7.1.transform_IUPAC_names_into_numerical_data.ipynb <==
# Transform IUPAC names into numerical data 

import pandas as pd

df_cid_smiles_iupac = pd.read_csv('CID_SMILES_iupac.csv', sep=';')

# Display the data frame
print('Shape of the data frame: ', df_cid_smiles_iupac.shape)
df_cid_smiles_iupac.head()

df_cid_smiles = pd.read_csv('input/smiles.csv')

# Display the data frame
print('Shape of the data frame: ', df_cid_smiles.shape)
df_cid_smiles.head()

df_cid_targets = pd.read_csv('input/targets.csv')

# Display the data frame
print('Shape of the data frame: ', df_cid_targets.shape)
df_cid_targets.head()

df_cid_smiles_targets = pd.merge( df_cid_smiles, df_cid_targets,on='SMILES')

# Display the data frame
print('Shape of the data frame: ', df_cid_smiles_targets.shape)
df_cid_smiles_targets.head()

df_cid_smiles_target_iupac = pd.merge(df_cid_smiles_iupac, df_cid_smiles_targets, on=['SMILES','CID'], how='inner')

# Display the data frame
print('Shape of the data frame: ', df_cid_smiles_target_iupac.shape)
df_cid_smiles_target_iupac.head()

df_cid_iupac_target = df_cid_smiles_target_iupac[['CID', 'IUPAC_name', 'target']]

# Display the data frame
print('Shape of the data frame: ', df_cid_iupac_target.shape)
df_cid_iupac_target.head()

# Remove the target column 
df_cids = df_cid_iupac_target.drop(columns=['target'], axis=1)

df_cids = df_cids[df_cids['IUPAC_name'].notna()]
# df = df.dropna()
df_cids['IUPAC_name'] = df_cids['IUPAC_name'].astype('string') 
df_cids.dtypes

# Obtain listo fo IUPAC formulas provided in PubChem
IUPAC_list = df_cids['IUPAC_name'].to_list()

for item in IUPAC_list[:5]:  # https://theprogrammingexpert.com/python-print-first-10-items-in-list/
    print(item)

from string import digits

# Remove elements from IUPAC formulas
dataset = []
for i in IUPAC_list:
    # Replace simbols with a coma
    new_list = i.translate(str.maketrans({'[': ' , ', ']': ' , ', 
                                         '(': ' , ', ')' : ' , ',
                                         '-': ' , ', ';' : ' , '}))
    
    new_str = new_list.translate(str.maketrans('', '', digits))
    dataset.append(new_str)
dataset

for item in dataset[:5]:  # https://theprogrammingexpert.com/python-print-first-10-items-in-list/
    print(item)

# Convert the strings into list of strings
new_strings = []
for i in dataset:
    parsed_string = i.split(",")
    new_strings.append(parsed_string)
new_strings

for item in new_strings[:5]:  # https://theprogrammingexpert.com/python-print-first-10-items-in-list/
    print(item)

from functools import reduce
# Convert list of lists in one list 
new_list = reduce(lambda a, b: a + b, new_strings)

for item in new_list[:5]:  # https://theprogrammingexpert.com/python-print-first-10-items-in-list/
    print(item)

# # Remove empty strings
# new_list = list(filter(None, new_list))
# new_list

new_list = [s.replace(' ', '') for s in new_list]
for item in new_list[:5]:  # https://theprogrammingexpert.com/python-print-first-10-items-in-list/
    print(item)

# Print the muber of left strings after reduction
print (len (new_list))

# remove duplicates
from collections import OrderedDict

new_list = list(OrderedDict.fromkeys(new_list)) 

for item in new_list[:5]:  # https://theprogrammingexpert.com/python-print-first-10-items-in-list/
    print(item)

# Remove one letter strings
new_list_columns = [x for x in new_list if len(x) >= 4]

for item in new_list_columns[:5]:   # https://theprogrammingexpert.com/python-print-first-10-items-in-list/
    print(item)

print (len (new_list_columns))

new_list_columns.sort()
new_list_columns

import numpy as np

dict_func_groups = dict.fromkeys(new_list_columns, 0)
dict_func_groups

# # check if a key is in dict
# 'acetamide' in dict_func_groups.keys()

# https://stackoverflow.com/questions/28056171/how-to-build-and-fill-pandas-dataframe-from-for-loop
import copy
import re
dataset = []
data = []
# new_strings = []

list_formulas = df_cids['IUPAC_name'].to_list()
for formula in list_formulas:
    new_list = re.findall(r'\b\w{4,}\b', formula)
    for group in new_list:
        # working
        new_dict = copy.deepcopy(dict_func_groups)
        for ii in new_list:
            if ii in new_dict.keys():
                new_dict[ii] +=1
                    
    id = formula
    # Combine all new features in an array
    # arr =[id, new_dict.values()]
    dataset.append(new_dict.values())

df_id = pd.DataFrame(list_formulas)

df_id  = df_id .rename(index=str, columns={0:'IUPAC_name'})

# Create an "index" column with the corresponding index values
df_id['index'] = df_id.index

# Convert the index values into strings
df_id['index'] = df_id['index'].astype(str)

# Display df_similarity
print('Shape of df_id: ', df_id.shape)
df_id.head() 

df_func_groups = pd.DataFrame.from_dict(dataset,
                                        orient='columns',
                                        dtype=None, columns=None) 


df_func_groups.columns = new_list_columns
df_func_groups.head()

# Create an "index" column with the corresponding index values
df_func_groups['index'] = df_func_groups.index

# Convert the index values into strings
df_func_groups['index'] = df_func_groups['index'].astype(str)

# Display df_similarity
print('Shape of df_func_groups: ', df_func_groups.shape)
df_func_groups.head() 

# Add the new data frame df_similarity to the already existing df based on "index"
df = pd.merge(df_id,df_func_groups, on=['index'])

# Display df_similarity
print('Shape of df: ', df.shape)
df.head()

df.drop('index', axis=1, inplace=True)
df.head()

# Merge the data frame with the targets 
df = pd.merge(df, df_cids, on='IUPAC_name')

# Display the data frame
print('Shape of df: ', df.shape)
df.head()

# # Converting object data type to numeric data type
# df['CID'] = pd.to_numeric(df['CID'])
# df['CID'].dtypes

df_IUPAC_targets = pd.merge( df_cid_iupac_target, df, on= ['CID', 'IUPAC_name'])

print('Shape of the data frame: ', df_IUPAC_targets.shape)
df_IUPAC_targets.head(2)

# Drop the CID and UPAC columns by name
df_IUPAC_targets = df_IUPAC_targets.drop(columns=['CID', 'IUPAC_name']) 

df_IUPAC_targets.to_csv('data_IUPACs.csv')
