==> 9.1.CHOP_create_4_csv_files_with_iupac_names.ipynb <==
# Create the csv files containing CID, SMILES notations and IUPAC names

import pandas as pd
import json

#Upload data
with open('json_CHOP.json', 'r', encoding='utf-8') as f:
    data = json.load(f)

# read json as a data frame 
df = pd.DataFrame.from_dict(pd.json_normalize(data))

# To avoid truncation of some columns during data frame display
pd.set_option('display.max_columns', None) 

# Display the data frame
print('Shape of the data frame: ', df.shape)
df.head()

df.info()

df = df[['cid', 'smiles', 'iupacname']]
df.head()

# Drop duplicates
df = df.drop_duplicates()
df.shape

df = df.dropna()
df.shape

# Rename column 'cid'
df = df.rename(index=str, columns={'cid': 'CID',
                                   'smiles' : 'SMILES',
                                   'iupacname':'UPAC'})
df.head()

# Split data frame into 7 parts
df1 = df.iloc[:55000]
df2 = df.iloc[55001:110000]
df3 = df.iloc[110001:165000]
df4 = df.iloc[165001:]

# Create csv file for each part of the data frame
df1.to_csv('df1.csv')
df2.to_csv('df2.csv')
df3.to_csv('df3.csv')
df4.to_csv('df4.csv')
