==> 9.2.CHOP_create_4_csv_files_with_targets.ipynb <==
# Create the csv files containing CID, SMILES notations and labels

import pandas as pd 

# load the BioAssay PubChem AID 2732 dataset
# https://pubchem.ncbi.nlm.nih.gov/bioassay/2732
df = pd.read_csv('input/pubchem_CHOP_bioassay.csv', sep=';')

# To avoid truncation of some columns during data frame display
pd.set_option('display.max_columns', None) 

# Display the data frame
print('Shape of the data frame: ', df.shape)

df.head()

df.info()

# Remove duplicates in general
df.drop_duplicates()
df.shape  # 359,035 means no duplicates

# Remove the compounds missing CID-s
df = df[df['CID'].notna()]
df.shape

# Turn CID float data type into integer
df = df.copy()
df['CID'] = df['CID'].astype('int64') 
df.info()

# Remove duplicates in CIDs 
df = df.drop_duplicates(subset=['CID'], keep=False) 
df.shape  # 6,446 are duplicates 

df = df [['CID', 'SMILES', 'target']]
df.head()

# Counting of values in column PUBCHEM_ACTIVITY_OUTCOME
df['target'].value_counts()

# Set the binary targets to "1" for an active inhibitors and "0" for others 
df = df.copy()

# Create a dictionary
target = {'Active' : 1,
          'Inactive' : 0}

# Replace the names with binary values set up in the dictionary targets
df.target = [target[item] for item in df.target]


# Check how many unique values there are in the 'target' column
df['target'].unique()

# Shuffle the resulting data sets
df = df.sample(
    frac = 1,        # Return entire dataframe
    random_state=1   # Make result reproducible
    ).reset_index(drop=True)

# Display the data frame
print('Shape of the data frame: ', df.shape)

df.head()

df.to_csv('CHOP_full_dataset_for_CID_list.csv')

[<a href="#content">Back to top</a>]

## Create csv files <a name="3"></a>

# First part of df
df_1 = pd.read_csv('df1.csv', index_col=[0])
df1= pd.merge(df, df_1, on=['CID', 'SMILES'])
print(df1.shape)
df1.head()

# Remain only the necessary columns
df1 = df1[['UPAC', 'target']] 
print(df1.shape)
df1.head()

# Generate the csv file
df1.to_csv('df1_targets.csv')

# First part of df
df_2 = pd.read_csv('df2.csv', index_col=[0])
df2= pd.merge(df, df_2, on=['CID', 'SMILES'])
print(df2.shape)
df2.head()

# Remain only the necessary columns
df2 = df2[['UPAC', 'target']] 
print(df2.shape)
df2.head()

# Generate the csv file
df2.to_csv('df2_targets.csv')

# Second part of df
df_3 = pd.read_csv('df3.csv', index_col=[0])
df3= pd.merge(df, df_3, on=['CID', 'SMILES'])
print(df3.shape)
df3.head()

# Remain only the necessary columns
df3 = df3[['UPAC', 'target']] 
print(df3.shape)
df3.head()

# Generate the csv file
df3.to_csv('df3_targets.csv')

# First part of df
df_4 = pd.read_csv('df4.csv', index_col=[0])
df4= pd.merge(df, df_4, on=['CID', 'SMILES'])
print(df4.shape)
df4.head()

# Remain only the necessary columns
df4 = df4[['UPAC', 'target']] 
print(df4.shape)
df4.head()

# Generate the csv file
df4.to_csv('df4_targets.csv')

[<a href="#content">Back to top</a>]
