==> 9.5.CHOP_group_all_dfs.ipynb <==
### Combine the resulted data frames with functional groups, the presence of a functional group in a compound and the relevent lable for the compound. 
### Display the functional groups that have participate in compoounds that have been identified as CHOP infibitors.

import pandas as pd
import numpy as np

# Upload data frames 
df1 = pd.read_csv('functional_groups_from_df1.csv', index_col=[0])
# df2 = pd.read_csv('functional_groups_from_df2.csv', index_col=[0]) empty df
# df3 = pd.read_csv('functional_groups_from_df3.csv', index_col=[0]) empty df
# df4 = pd.read_csv('functional_groups_from_df4.csv', index_col=[0]) emprty df

# #Create a list with the data frames 
# list_of_dataframes = [df1, df2, df3, df4]

# # Concatenate the data frames 
# combined_df = pd.concat(list_of_dataframes, ignore_index=True)

# Display the data frame
pd.set_option('display.max_rows', None)  # display all rows
print(df1.shape)
df1.head(267)

df1 = pd.read_csv('df1_for_final.csv', index_col=[0])
# df2 = pd.read_csv('df2_for_final.csv', index_col=[0])
# df3 = pd.read_csv('df3_for_final.csv', index_col=[0])
# df4 = pd.read_csv('df4_for_final.csv', index_col=[0])

# # Create a list with the data frames 
# df = [df1, df2, df3, df4]

# # Concatenate the data frames 
# df = pd.concat(df, ignore_index=True)

# Display the data frame
print(df1.shape)
df1.head()

# Define the columns and the condition
column_to_get_value = 'UPAC'              # The column from which the values will be retreaved 
column_with_condition = 'hexahydroazuleno'  # The column to check if its value is 1
condition_value = 1                       # The specific value (1) 

# Filter the DataFrame and select the desired column
# This operation first filters the rows where 'is_active' is 1,
# Then selects the 'item_name' from those filtered rows.
result_series = df1[df1[column_with_condition] == condition_value][column_to_get_value]

# Print the result(s)
# print(f"Values in '{column_to_get_value}' where '{column_with_condition}' is '{condition_value}':")

if not result_series.empty:
    for value in result_series:
        print(value)
    # Alternatively, print as a list:
        # print(result_series.tolist())
else:
    print("No rows found where the condition is met.")

print("-" * 40)
